+++ src/libopensc/Makefile.am
@@ -85,7 +85,8 @@ libopensc_la_LIBADD = $(OPENPACE_LIBS) $(OPTIONAL_OPENSSL_LIBS) \
 	$(top_builddir)/src/common/libscdl.la \
 	$(top_builddir)/src/ui/libnotify.la \
 	$(top_builddir)/src/ui/libstrings.la \
-	$(top_builddir)/src/sm/libsmeac.la
+	$(top_builddir)/src/sm/libsmeac.la \
+	$(top_builddir)/src/sm/libsmjacartapki.la
 if WIN32
 libopensc_la_LIBADD += -lws2_32 -lshlwapi -lcomctl32
 endif
