+++ src/libopensc/Makefile.mak
@@ -46,8 +46,10 @@ LIBS = $(TOPDIR)\src\scconf\scconf.lib \
 	   $(TOPDIR)\src\common\libscdl.lib \
 	   $(TOPDIR)\src\ui\strings.lib \
 	   $(TOPDIR)\src\ui\notify.lib \
+	   $(TOPDIR)\src\sm\libsm.lib \
 	   $(TOPDIR)\src\sm\libsmiso.lib \
 	   $(TOPDIR)\src\sm\libsmeac.lib \
+	   $(TOPDIR)\src\sm\libsmjacartapki.lib \
 	   $(TOPDIR)\src\pkcs15init\pkcs15init.lib
 
 all: $(TOPDIR)\win32\versioninfo.res $(TARGET)
