+++ src/sm/Makefile.am
@@ -3,12 +3,13 @@
 MAINTAINERCLEANFILES = Makefile.in
 EXTRA_DIST = Makefile.mak
 
-noinst_LTLIBRARIES = libsmiso.la libsmeac.la
+noinst_LTLIBRARIES = libsmiso.la libsmeac.la libsmjacartapki.la
 
 noinst_HEADERS = \
 		 sm-iso-internal.h \
 		 sm-iso.h \
-		 sm-eac.h
+		 sm-eac.h \
+		 sm-jacartapki.h
 
 if ENABLE_OPENSSL
 noinst_LTLIBRARIES += libsm.la
@@ -27,3 +28,7 @@ libsmiso_la_SOURCES = sm-iso.c
 libsmeac_la_SOURCES = sm-eac.c
 libsmeac_la_LIBADD = $(OPENPACE_LIBS) $(OPENSSL_LIBS) libsmiso.la
 libsmeac_la_CFLAGS = $(OPENPACE_CFLAGS) $(OPENSSL_CFLAGS) -I$(top_srcdir)/src
+
+libsmjacartapki_la_SOURCES = sm-jacartapki.c
+libsmjacartapki_la_LIBADD = $(OPENSSL_LIBS) libsm.la
+libsmjacartapki_la_CFLAGS = $(OPENSSL_CFLAGS) -I$(top_srcdir)/src
