+++ src/sm/Makefile.mak
@@ -9,7 +9,10 @@ OBJECTS1 = sm-iso.obj
 TARGET2 = libsmeac.lib
 OBJECTS2 = sm-eac.obj
 
-all: $(TARGET) $(TARGET1) $(TARGET2)
+TARGET3 = libsmjacartapki.lib
+OBJECTS3 = sm-jacartapki.obj
+
+all: $(TARGET) $(TARGET1) $(TARGET2) $(TARGET3)
 
 !INCLUDE $(TOPDIR)\win32\Make.rules.mak
 
@@ -28,3 +31,6 @@ $(TARGET1): $(OBJECTS1)
 
 $(TARGET2): $(OBJECTS2)
         lib $(LIBFLAGS) /out:$(TARGET2) $(OBJECTS2)
+
+$(TARGET3): $(OBJECTS3)
+        lib $(LIBFLAGS) /out:$(TARGET3) $(OBJECTS3)
